Version 5
Made significant changes on the Dashboard, and other user profiles!

+++ database/migrations/2018_03_08_003456_create_job__type_table.php
@@ -18,7 +18,7 @@ class CreateJobTypeTable extends Migration
             $table->string('type');
             $table->string('description');
 
-            $table->mediumInteger('specs_id')->nullable()->unsigned();
+            $table->integer('specs_id')->nullable()->unsigned();
             $table->timestamps();
         });
     }
